requirements.txt
# Web Automation Dependencies
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# AI Integration
google-generativeai==0.3.2

# Firebase Integration
firebase-admin==6.3.0
# Environment Management
python-dotenv==1.0.0

# Data Processing
requests==2.31.0

# Optional: For enhanced functionality
Pillow==10.1.0
pandas==2.1.4